refactor: renderWithTheme 함수 제작
