+++ README.md
@@ -15,5 +15,6 @@
 - Form Field 컴포넌트 구현
 - UnderlineInputField를 통해 Login 페이지 리팩토링
 - Typography 테스트 코드 작성
+- renderWithTheme 함수 제작후 리팩토링
 
 ## 리뷰 반영
