+++ src/components/common/Typography/__test__/Typography.test.tsx
@@ -1,24 +1,16 @@
-import { render, screen } from '@testing-library/react';
+import { screen } from '@testing-library/react';
 import { Typography } from '../Typography';
-import { ThemeProvider } from '@emotion/react';
-import { theme } from '@styles/tokens';
+
+import { renderWithTheme } from '@test/utils/renderWithTheme';
 
 describe('Typography', () => {
   it('renders children', () => {
-    render(
-      <ThemeProvider theme={theme}>
-        <Typography>텍스트</Typography>
-      </ThemeProvider>
-    );
+    renderWithTheme(<Typography>텍스트</Typography>);
     expect(screen.getByText('텍스트')).toBeInTheDocument();
   });
 
   it('HTML 태그 테스트', () => {
-    render(
-      <ThemeProvider theme={theme}>
-        <Typography as="h2">제목</Typography>
-      </ThemeProvider>
-    );
+    renderWithTheme(<Typography as="h2">제목</Typography>);
     const el = screen.getByText('제목');
     expect(el.tagName).toBe('H2');
   });
